PR1: extract studio_core.services.dashboard + CLI shims
Move UI-agnostic build_studio_dashboard() to studio_core so CLI/TUI/Web
share one snapshot contract. tools/cli/dashboard.py re-exports the builder
and keeps Rich rendering. Adds path setup, docs, and tests.
